requirements.txt
pyserial==3.5
opencv-python==4.9.0.80
numpy==1.26.4
